# swagger.yaml
openapi: 3.0.0
info:
  title: Node.js TypeScript MySQL Boilerplate API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /accounts/register:
    post:
      summary: Register a new account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string }
                password: { type: string }
                confirmPassword: { type: string }
                acceptTerms: { type: boolean }
      responses:
        '200': { description: Registration successful }
  /accounts/authenticate:
    post:
      summary: Authenticate and get JWT
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Authentication successful }
  /accounts/verify-email:
    post:
      summary: Verify email with token
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        '200': { description: Verification successful }
  /accounts/forgot-password:
    post:
      summary: Send password reset email
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200': { description: Email sent }
  /accounts/reset-password:
    post:
      summary: Reset password with token
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                password: { type: string }
                confirmPassword: { type: string }
      responses:
        '200': { description: Password reset successful }